alias(libs.plugins.aboutlibraries) is actually needed after all to generate the json file
