+++ shared/feature/licenses/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     id("ly.david.android.library")
     id("ly.david.musicsearch.compose.multiplatform")
     id("ly.david.musicsearch.kotlin.multiplatform")
+    alias(libs.plugins.aboutlibraries)
 }
 
 android {
@@ -16,11 +17,9 @@ kotlin {
                 implementation(projects.strings)
                 implementation(projects.shared.screens)
 
-                implementation(compose.foundation)
-                implementation(compose.materialIconsExtended)
+//                implementation(compose.foundation)
                 implementation(compose.material3)
-                implementation(compose.ui)
-                implementation(compose.preview)
+//                implementation(compose.ui)
 
                 implementation(libs.aboutlibraries.compose)
                 implementation(libs.aboutlibraries.core)
